restructured my conversions file and now conversions are working for types d and i. Placeholder value types are also in types u, x, X, o so gotta make sure to update those

+++ libft/ft_itoa.c
@@ -6,7 +6,7 @@
 /*   By: malasalm <malasalm@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/24 16:39:48 by malasalm          #+#    #+#             */
-/*   Updated: 2019/11/10 16:22:43 by malasalm         ###   ########.fr       */
+/*   Updated: 2020/08/07 18:10:14 by malasalm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,11 +19,11 @@
 
 #include "libft.h"
 
-char	*ft_itoa(int n)
+char	*ft_itoa(long long n)
 {
-	char	*str;
-	long	num;
-	int		i;
+	char		*str;
+	long long	num;
+	int			i;
 
 	num = n;
 	i = ft_intlen(num);
